+++ src/util/pix-png.c
@@ -38,16 +38,19 @@ static const char RCSid[] = "@(#)$Header$ (BRL)";
 
 #include <stdio.h>
 #include <math.h>
+
 #include "zlib.h"
 #include "pngconf.h"
 #include "png.h"
+
 #include "machine.h"
 #include "bu.h"
 #include "vmath.h"
 #include "bn.h"
 
-static int	file_width = 512;		/* default input width */
-static int	file_height = 512;		/* default input height */
+
+static long int	file_width = 512L;		/* default input width */
+static long int	file_height = 512L;		/* default input height */
 static int	autosize = 0;			/* !0 to autosize input */
 static int	fileinput = 0;			/* file of pipe on input? */
 static char	*file_name;
@@ -74,15 +77,15 @@ get_args(int argc, register char **argv)
 			break;
 		case 's':
 			/* square file size */
-			file_height = file_width = atoi(optarg);
+			file_height = file_width = atol(optarg);
 			autosize = 0;
 			break;
 		case 'w':
-			file_width = atoi(optarg);
+			file_width = atol(optarg);
 			autosize = 0;
 			break;
 		case 'n':
-			file_height = atoi(optarg);
+			file_height = atol(optarg);
 			autosize = 0;
 			break;
 
@@ -130,10 +133,10 @@ main(int argc, char **argv)
 
 	/* autosize input? */
 	if( fileinput && autosize ) {
-		int	w, h;
+		unsigned long int	w, h;
 		if( bn_common_file_size(&w, &h, file_name, 3) ) {
-			file_width = w;
-			file_height = h;
+			file_width = (long)w;
+			file_height = (long)h;
 		} else {
 			fprintf(stderr,"pix-png: unable to autosize\n");
 		}
